(refactor) Refactor mongo db to have setup and index file
